requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-multipart==0.0.20
email-validator==2.3.0

 # LangChain + LLMs
langchain==0.3.7
langchain-community==0.3.7
langchain-google-genai==2.0.1

 # Embeddings
 sentence-transformers==3.0.1

 # Supabase and DB
 supabase==2.6.0
 psycopg[binary]==3.2.1
 pgvector==0.3.3
 SQLAlchemy==2.0.35

# Resume parsing
# Note: pyresparser has compatibility issues, using alternative approach
# pyresparser==1.0.6
# spacy==3.7.5
# nltk==3.9.1
pdfplumber==0.11.0
PyPDF2==3.0.1
python-docx==1.1.2

 # Utilities
 numpy==1.26.4
 scikit-learn==1.5.2


